+++ src/control/funtzioak/FuntzioBista.java
@@ -12,6 +12,12 @@ import model.objektuak.Podcast;
 import model.objektuak.Podcaster;
 import vista.*;
 import vista.bezeroa.*;
+import vista.bezeroa.musika.AlbumDeskubritu;
+import vista.bezeroa.musika.MusikaDeskubritu;
+import vista.bezeroa.musika.MusikaIkusi;
+import vista.bezeroa.playlist.NirePlaylist;
+import vista.bezeroa.podcast.PodcastDeskubritu;
+import vista.bezeroa.podcast.PodcastIkusi;
 
 
 /**
