+++ src/vista/bezeroa/musika/MusikaIkusi.java
@@ -1,4 +1,4 @@
-package vista.bezeroa;
+package vista.bezeroa.musika;
 
 import java.awt.BorderLayout;
 import java.awt.Font;
@@ -23,8 +23,10 @@ import javax.swing.table.DefaultTableModel;
 import control.funtzioak.FuntzioBista;
 import control.funtzioak.Funtzioak;
 import model.Aldagaiak;
+import model.dao.AbestiaDao;
 import model.dao.AlbumDao;
 import model.dao.PodcastDao;
+import model.objektuak.Musikaria;
 
 import java.awt.GridLayout;
 import javax.swing.JTable;
@@ -33,20 +35,21 @@ import javax.swing.ScrollPaneConstants;
 
 import model.objektuak.*;
 
-public class PodcastIkusi extends JFrame {
+public class MusikaIkusi extends JFrame {
 
 	private static final long serialVersionUID = 1L;
 	private JPanel contentPane;
 	private JTable table;
 	private DefaultTableModel model;
-	PodcastDao podcastDao = new PodcastDao();
-	ArrayList <Audio> podcastList;
-	
+	AbestiaDao abestiDao = new AbestiaDao();
+	ArrayList<Audio> abestiList;
+
 	/**
 	 * Create the frame.
-	 * @throws SQLException 
+	 * 
+	 * @throws SQLException
 	 */
-	public PodcastIkusi(Podcaster podcaster) throws SQLException {
+	public MusikaIkusi(Musikaria musikaria, Album album) throws SQLException {
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		setIconImage(Toolkit.getDefaultToolkit().getImage(MusikaDeskubritu.class.getResource(Aldagaiak.logo)));
 		setBounds(Aldagaiak.cordX, Aldagaiak.cordY, Aldagaiak.resolucionX, Aldagaiak.resolucionY);
@@ -55,16 +58,13 @@ public class PodcastIkusi extends JFrame {
 
 		setContentPane(contentPane);
 		contentPane.setLayout(new BorderLayout(0, 0));
-		
-		podcastList = new ArrayList<Audio>();
-		podcastList = podcastDao.getPodcastByPodcasterId(podcaster);
-		
 
-		
+		abestiList = new ArrayList<Audio>();
+		abestiList = abestiDao.getAbestiaByAlbumId(album);
+
 		JPanel panelHeader = new JPanel();
 		contentPane.add(panelHeader, BorderLayout.NORTH);
 		panelHeader.setLayout(new BorderLayout(0, 0));
-		
 
 		// Erabiltzailearen izena bistaratzeko botoia
 		JButton btnPerfil = new JButton(Aldagaiak.erabiltzailea.getErabiltzaileIzena());
@@ -84,104 +84,102 @@ public class PodcastIkusi extends JFrame {
 		btnAtzera.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				FuntzioBista.bistaAldatu(getBounds(), getWidth(), getHeight());
-				FuntzioBista.irekiPodcastDeskubritu();
+				FuntzioBista.irekiAlbumDeskubritu(musikaria);
 				dispose();
 			}
 		});
 		panelHeader.add(btnAtzera, BorderLayout.WEST);
-		
-		JLabel lblPodcastIzena = new JLabel(podcaster.getIzen_Artistikoa() + "-aren podcastak");
-		lblPodcastIzena.setFont(new Font("Tahoma", Font.PLAIN, 35));
-		lblPodcastIzena.setHorizontalAlignment(SwingConstants.CENTER);
-		panelHeader.add(lblPodcastIzena, BorderLayout.CENTER);
-		
+
+		JLabel lblAlbumIzena = new JLabel(album.getIzenburua() + "-aren musika");
+		lblAlbumIzena.setFont(new Font("Tahoma", Font.PLAIN, 35));
+		lblAlbumIzena.setHorizontalAlignment(SwingConstants.CENTER);
+		panelHeader.add(lblAlbumIzena, BorderLayout.CENTER);
+
 		JPanel panel_1 = new JPanel();
 		contentPane.add(panel_1, BorderLayout.CENTER);
 		panel_1.setLayout(new GridLayout(0, 2, 0, 0));
-		
+
 		JPanel panelPodcastTabla = new JPanel();
 		panel_1.add(panelPodcastTabla);
 		panelPodcastTabla.setLayout(new BorderLayout(0, 0));
-		
+
+		System.out.println(abestiList);
+
 		JLabel lblNewLabel_1 = new JLabel(" ");
 		panelPodcastTabla.add(lblNewLabel_1, BorderLayout.NORTH);
-		
+
 		JLabel lblNewLabel_2 = new JLabel(" ");
 		panelPodcastTabla.add(lblNewLabel_2, BorderLayout.SOUTH);
-		
+
 		JLabel lblNewLabel_3 = new JLabel("      ");
 		panelPodcastTabla.add(lblNewLabel_3, BorderLayout.WEST);
-		
+
 		JLabel lblNewLabel_4 = new JLabel("      ");
 		panelPodcastTabla.add(lblNewLabel_4, BorderLayout.EAST);
-		
+
 		JPanel panel_3 = new JPanel();
 		panel_1.add(panel_3);
 		panel_3.setLayout(new GridLayout(0, 1, 0, 0));
-		
+
 		JPanel panel_4 = new JPanel();
 		panel_3.add(panel_4);
-		
 		panel_4.setLayout(new BorderLayout(0, 0));
-		
+
 		JTextPane textPaneDeskripzioa = new JTextPane();
 		textPaneDeskripzioa.setEditable(false);
 		textPaneDeskripzioa.setFocusable(false);
-		
-		textPaneDeskripzioa.setText("Deskribapena: " + podcaster.getDeskribapena());
+
+		textPaneDeskripzioa.setText("Deskribapena: " + album.getDeskripzioa());
+
 		panel_4.add(textPaneDeskripzioa, BorderLayout.CENTER);
-		
+
 		JLabel lblNewLabel_5 = new JLabel(" ");
 		panel_4.add(lblNewLabel_5, BorderLayout.NORTH);
-		
+
 		JLabel lblNewLabel_6 = new JLabel(" ");
 		panel_4.add(lblNewLabel_6, BorderLayout.SOUTH);
-		
+
 		JLabel lblNewLabel_7 = new JLabel("      ");
 		panel_4.add(lblNewLabel_7, BorderLayout.WEST);
-		
+
 		JLabel lblNewLabel_8 = new JLabel("      ");
 		panel_4.add(lblNewLabel_8, BorderLayout.EAST);
-		
+
 		JPanel panel_5 = new JPanel();
 		panel_3.add(panel_5);
 		panel_5.setLayout(new BorderLayout(0, 0));
-		
-		
+
 		ImageIcon icon = null;
-		try {
-			icon = new ImageIcon(podcaster.getIrudia().getBytes(1, (int) podcaster.getIrudia().length()));
-		} catch (SQLException e1) {
 		
+		try {
+			icon = new ImageIcon(album.getIrudia().getBytes(1, (int) album.getIrudia().length()));
+		} catch (SQLException e1) {
 			e1.printStackTrace();
 		}
-		
-		
+
 		JLabel lblIrudia = new JLabel("");
 		lblIrudia.setHorizontalAlignment(SwingConstants.CENTER);
-		// lblIrudia.setIcon(icon);
+		lblIrudia.setIcon(icon);
 		panel_5.add(lblIrudia, BorderLayout.CENTER);
-		
+
 		JLabel lblNewLabel_10 = new JLabel(" ");
 		panel_5.add(lblNewLabel_10, BorderLayout.NORTH);
-		
+
 		JLabel lblNewLabel_11 = new JLabel("      ");
 		panel_5.add(lblNewLabel_11, BorderLayout.WEST);
-		
+
 		JLabel lblNewLabel_12 = new JLabel(" ");
 		panel_5.add(lblNewLabel_12, BorderLayout.SOUTH);
-		
+
 		JLabel lblNewLabel_13 = new JLabel("      ");
 		panel_5.add(lblNewLabel_13, BorderLayout.EAST);
-		
-		/*----------------------------------------------------------------------------------------------------------------*/
-		
+
 		JScrollPane scrollPane = new JScrollPane();
 		scrollPane.setVerticalScrollBarPolicy(ScrollPaneConstants.VERTICAL_SCROLLBAR_NEVER);
 		panelPodcastTabla.add(scrollPane, BorderLayout.CENTER);
-		
-		String[] stringAux = {"PODCAST"};
-		
+
+		String[] stringAux = { "ABESTIAK" };
+
 		model = new DefaultTableModel();
 		table = new JTable(model);
 		table.setFont(new Font("Tahoma", Font.PLAIN, 20));
@@ -190,26 +188,26 @@ public class PodcastIkusi extends JFrame {
 		table.getTableHeader().setReorderingAllowed(false);
 		table.addMouseListener(new MouseAdapter() {
 			@Override
-			
+
 			public void mouseClicked(MouseEvent e) {
 				int index = table.getSelectedRow();
-				
+
 				FuntzioBista.bistaAldatu(getBounds(), getWidth(), getHeight());
-				
-				FuntzioBista.irekiErreprodukzioa(podcastList, index);
+
+				FuntzioBista.irekiErreprodukzioa(abestiList, index);
 				dispose();
 			}
-			
+
 		});
 		model.setColumnIdentifiers(stringAux);
-			
+
 		scrollPane.setViewportView(table);
 		Object[] aux = new Object[1];
-		for (int i = 0 ; i < podcastList.size(); i++) {
-			String iraupena = Funtzioak.secondsToString(podcastList.get(i).getIraupena());
-			aux[0] = podcastList.get(i).getIzena() + " (" + iraupena + ")";
-			
-            model.addRow(aux);
-        }
+		for (int i = 0; i < abestiList.size(); i++) {
+			String iraupena = Funtzioak.secondsToString(abestiList.get(i).getIraupena());
+			aux[0] = abestiList.get(i).getIzena() + " (" + iraupena + ")";
+
+			model.addRow(aux);
+		}
 	}
 }
